pyproject.toml
[tool.poetry]
name = "bayes-abund"
version = "0.1.0"
description = ""
authors = ["synedraacus <alexeymorozov1991@gmail.com>"]
readme = "README.md"
packages = [{include = "bayes_abund"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.4"
biopython = "^1.83"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "^23.12.1"
ruff = "^0.1.13"

[tool.ruff]
select = ["E", "F", "W", "I"]
fixable = ["ALL"]
fix = true



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
